feat(1251-show-lambda-logs): add lambda logs to UI

+++ src/db/include/awsmock/repository/LambdaDatabase.h
@@ -30,9 +30,9 @@ namespace AwsMock::Database {
     using std::chrono::system_clock;
 
     struct LambdaMonitoringCounter {
-        int instances{};
-        int invocations{};
-        int averageRuntime{};
+        long instances{};
+        long invocations{};
+        long averageRuntime{};
         system_clock::time_point modified = system_clock::now();
     };
